# openapi.yaml
openapi: 3.1.0
info:
  title: API de Escalas da Igreja
  version: 1.0.0
  description: |
    API RESTful para gerenciamento de celebrações, disponibilidades e geração de escalas.
servers:
  - url: https://your-domain.com
paths:
...
  /api/availabilities:
    post:
      summary: Criar ou atualizar disponibilidade
      description: >
        Marca a disponibilidade do membro logado para uma celebração.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                celebration_id:
                  type: string
                available:
                  type: boolean
              required: [celebration_id, available]
      responses:
        '200':
          description: Disponibilidade registrada
        '400':
          description: Erro na solicitação
        '401':
          description: Não autenticado
...